Renamed to 'image_generation'
