+++ .github/labeler.yml
@@ -117,10 +117,10 @@
 - 'pyproject.toml'
 
 'category: text to image':
-- 'src/include/openvino/genai/text2image/**/*'
-- 'src/cpp/src/text2image/**/*'
-- 'src/python/py_text2image_models.cpp'
-- 'src/python/py_text2image_pipeline.cpp'
+- 'src/include/openvino/genai/image_generation/**/*'
+- 'src/cpp/src/image_generation/**/*'
+- 'src/python/py_image_generation_models.cpp'
+- 'src/python/py_image_generation_pipelines.cpp'
 
 'category: GenAI C++ API':
 - 'src/cpp/include/openvino/genai/**/*'
