+++ .github/workflows/stable_diffusion_1_5_cpp.yml
@@ -77,7 +77,7 @@ jobs:
         run: |
           source ${{ env.OV_INSTALL_DIR }}/setupvars.sh
           ${{ env.build_dir }}/samples/cpp/text2image/lora_stable_diffusion ./models/dreamlike-art-dreamlike-anime-1.0/FP16 "curly-haired unicorn in the forest, anime, line" ./models/soulcard.safetensors 0.7
-      
+
       - name: Run Python main app
         run: |
           source openvino_sd_cpp/bin/activate
@@ -85,7 +85,7 @@ jobs:
           python ./samples/python/text2image/main.py ./models/dreamlike-art-dreamlike-anime-1.0/FP16 "cyberpunk cityscape like Tokyo New York with tall buildings at dusk golden hour cinematic lighting"
         env:
           PYTHONPATH: ${{ env.build_dir }}
-      
+
       - name: Run Python LoRA app
         run: |
           source openvino_sd_cpp/bin/activate
@@ -120,7 +120,7 @@ jobs:
         with:
           python-version: ${{ env.PYTHON_VERSION }}
           cache: 'pip'
-  
+
       - name: Build app
         run: |
           . "${{ env.OV_INSTALL_DIR }}/setupvars.ps1"
@@ -148,7 +148,7 @@ jobs:
           "${{ env.build_dir }}/samples/cpp/text2image/Release/stable_diffusion.exe ./models/dreamlike-art-dreamlike-anime-1.0/FP16 'cyberpunk cityscape like Tokyo New York with tall buildings at dusk golden hour cinematic lighting'"
         env:
           PATH: ${{ env.build_dir }}\openvino_genai
-  
+
       - name: Run LoRA app
         run: |
           . "${{ env.OV_INSTALL_DIR }}/setupvars.ps1"
