+++ src/cpp/include/openvino/genai/image_generation/autoencoder_kl.hpp
