+++ src/cpp/include/openvino/genai/image_generation/clip_text_model.hpp
