+++ src/cpp/include/openvino/genai/image_generation/clip_text_model_with_projection.hpp
