+++ src/cpp/include/openvino/genai/image_generation/text2image_pipeline.hpp
@@ -9,40 +9,16 @@
 #include <filesystem>
 
 #include "openvino/core/any.hpp"
-#include "openvino/runtime/properties.hpp"
 #include "openvino/runtime/tensor.hpp"
 
-#include "openvino/genai/visibility.hpp"
-
-#include "openvino/genai/lora_adapter.hpp"
-#include "openvino/genai/text2image/clip_text_model.hpp"
-#include "openvino/genai/text2image/clip_text_model_with_projection.hpp"
-#include "openvino/genai/text2image/unet2d_condition_model.hpp"
-#include "openvino/genai/text2image/autoencoder_kl.hpp"
+#include "openvino/genai/image_generation/clip_text_model.hpp"
+#include "openvino/genai/image_generation/clip_text_model_with_projection.hpp"
+#include "openvino/genai/image_generation/unet2d_condition_model.hpp"
+#include "openvino/genai/image_generation/autoencoder_kl.hpp"
+#include "openvino/genai/image_generation/generation_config.hpp"
 
 namespace ov {
 namespace genai {
-
-//
-// Random generators
-//
-
-class OPENVINO_GENAI_EXPORTS Generator {
-public:
-    virtual float next() = 0;
-    virtual ~Generator();
-};
-
-class OPENVINO_GENAI_EXPORTS CppStdGenerator : public Generator {
-public:
-    // creates 'std::mt19937' with initial 'seed' to generate numbers within a range [0.0f, 1.0f]
-    explicit CppStdGenerator(uint32_t seed);
-
-    virtual float next() override;
-private:
-    std::mt19937 gen;
-    std::normal_distribution<float> normal;
-};
 
 //
 // Text to image pipeline
@@ -64,41 +40,6 @@ public:
                                                       Type scheduler_type = AUTO);
 
         virtual ~Scheduler();
-    };
-
-    struct OPENVINO_GENAI_EXPORTS GenerationConfig {
-        // LCM: prompt only w/o negative prompt
-        // SD XL: prompt2 and negative_prompt2
-        // FLUX: prompt2 (prompt if prompt2 is not defined explicitly)
-        // SD 3: prompt2, prompt3 (with fallback to prompt) and negative_prompt2, negative_prompt3
-        std::optional<std::string> prompt_2 = std::nullopt, prompt_3 = std::nullopt;
-        std::string negative_prompt, negative_prompt_2, negative_prompt_3;
-
-        // Optional image for image to image generation
-        ov::Tensor image;
-
-        size_t num_images_per_prompt = 1;
-
-        // random generator to have deterministic results
-        std::shared_ptr<Generator> random_generator = std::make_shared<CppStdGenerator>(42);
-
-        // the following values depend on HF diffusers class used to perform generation
-        float guidance_scale = 7.5f;
-        int64_t height = -1;
-        int64_t width = -1;
-        size_t num_inference_steps = 50;
-
-        AdapterConfig adapters;
-
-        void update_generation_config(const ov::AnyMap& config_map);
-
-        // checks whether is config is valid
-        void validate() const;
-
-        template <typename... Properties>
-        ov::util::EnableIfAllStringAny<void, Properties...> update_generation_config(Properties&&... properties) {
-            return update_generation_config(ov::AnyMap{std::forward<Properties>(properties)...});
-        }
     };
 
     explicit Text2ImagePipeline(const std::filesystem::path& models_path);
@@ -134,8 +75,8 @@ public:
         const UNet2DConditionModel& unet,
         const AutoencoderKL& vae);
 
-    GenerationConfig get_generation_config() const;
-    void set_generation_config(const GenerationConfig& generation_config);
+    ImageGenerationConfig get_generation_config() const;
+    void set_generation_config(const ImageGenerationConfig& generation_config);
 
     // ability to override scheduler
     void set_scheduler(std::shared_ptr<Scheduler> scheduler);
@@ -164,30 +105,6 @@ private:
     class StableDiffusionPipeline;
     class StableDiffusionXLPipeline;
 };
-
-//
-// Generation config properties
-//
-
-static constexpr ov::Property<std::string> prompt_2{"prompt_2"};
-static constexpr ov::Property<std::string> prompt_3{"prompt_3"};
-
-static constexpr ov::Property<std::string> negative_prompt{"negative_prompt"};
-static constexpr ov::Property<std::string> negative_prompt_2{"negative_prompt_2"};
-static constexpr ov::Property<std::string> negative_prompt_3{"negative_prompt_3"};
-
-static constexpr ov::Property<ov::Tensor> image{"image"};
-
-static constexpr ov::Property<size_t> num_images_per_prompt{"num_images_per_prompt"};
-static constexpr ov::Property<float> guidance_scale{"guidance_scale"};
-static constexpr ov::Property<int64_t> height{"height"};
-static constexpr ov::Property<int64_t> width{"width"};
-static constexpr ov::Property<size_t> num_inference_steps{"num_inference_steps"};
-
-static constexpr ov::Property<std::shared_ptr<Generator>> random_generator{"random_generator"};
-
-OPENVINO_GENAI_EXPORTS
-std::pair<std::string, ov::Any> generation_config(const Text2ImagePipeline::GenerationConfig& generation_config);
 
 } // namespace genai
 } // namespace ov
