+++ src/cpp/include/openvino/genai/image_generation/unet2d_condition_model.hpp
