+++ src/cpp/src/image_generation/models/clip_text_model_with_projection.cpp
@@ -1,7 +1,7 @@
 // Copyright (C) 2023-2024 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-#include "openvino/genai/text2image/clip_text_model_with_projection.hpp"
+#include "openvino/genai/image_generation/clip_text_model_with_projection.hpp"
 
 #include <fstream>
