+++ src/cpp/src/image_generation/numpy_utils.hpp
