+++ src/cpp/src/image_generation/schedulers/ddim.hpp
@@ -4,43 +4,36 @@
 #pragma once
 
 #include <filesystem>
+#include <list>
 #include <string>
-#include <random>
-#include <vector>
 
-#include "text2image/schedulers/types.hpp"
-#include "text2image/schedulers/ischeduler.hpp"
+#include "image_generation/schedulers/types.hpp"
+#include "image_generation/schedulers/ischeduler.hpp"
 
 namespace ov {
 namespace genai {
 
-class LCMScheduler : public IScheduler {
+class DDIMScheduler : public IScheduler {
 public:
-    // values from https://github.com/huggingface/diffusers/blob/main/src/diffusers/schedulers/scheduling_lcm.py#L190
     struct Config {
-        size_t num_train_timesteps = 1000;
-        float beta_start = 0.00085f, beta_end = 0.012f;
+        int32_t num_train_timesteps = 1000;
+        float beta_start = 0.0001f, beta_end = 0.02f;
         BetaSchedule beta_schedule = BetaSchedule::SCALED_LINEAR;
         std::vector<float> trained_betas = {};
-        size_t original_inference_steps = 50;
-        bool clip_sample = false;
-        float clip_sample_range = 1.0f;
-        bool set_alpha_to_one = true;
+        bool clip_sample = true, set_alpha_to_one = true;
         size_t steps_offset = 0;
         PredictionType prediction_type = PredictionType::EPSILON;
         bool thresholding = false;
-        float dynamic_thresholding_ratio = 0.995f;
-        float sample_max_value = 1.0f;
+        float dynamic_thresholding_ratio = 0.995f, clip_sample_range = 1.0f, sample_max_value = 1.0f;
         TimestepSpacing timestep_spacing = TimestepSpacing::LEADING;
-        float timestep_scaling = 10.0f;
         bool rescale_betas_zero_snr = false;
 
         Config() = default;
         explicit Config(const std::filesystem::path& scheduler_config_path);
     };
 
-    explicit LCMScheduler(const std::filesystem::path& scheduler_config_path);
-    explicit LCMScheduler(const Config& scheduler_config);
+    explicit DDIMScheduler(const std::filesystem::path& scheduler_config_path);
+    explicit DDIMScheduler(const Config& scheduler_config);
 
     void set_timesteps(size_t num_inference_steps, float strength) override;
 
@@ -59,16 +52,9 @@ private:
 
     std::vector<float> m_alphas_cumprod;
     float m_final_alpha_cumprod;
+
     size_t m_num_inference_steps;
-    float m_sigma_data;
-
     std::vector<int64_t> m_timesteps;
-
-    uint32_t m_seed;
-    std::mt19937 m_gen;
-    std::normal_distribution<float> m_normal;
-
-    std::vector<float> threshold_sample(const std::vector<float>& flat_sample);
 };
 
 } // namespace genai
