+++ src/cpp/src/image_generation/schedulers/types.cpp
@@ -1,7 +1,7 @@
 // Copyright (C) 2023-2024 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-#include "text2image/schedulers/types.hpp"
+#include "image_generation/schedulers/types.hpp"
 
 namespace ov {
 namespace genai {
