+++ src/python/py_openvino_genai.cpp
@@ -29,7 +29,7 @@ void init_generation_config(py::module_& m);
 
 void init_continuous_batching_pipeline(py::module_& m);
 void init_llm_pipeline(py::module_& m);
-void init_text2image_pipeline(py::module_& m);
+void init_image_generation_pipelines(py::module_& m);
 void init_vlm_pipeline(py::module_& m);
 void init_whisper_pipeline(py::module_& m);
 
@@ -117,7 +117,7 @@ PYBIND11_MODULE(py_openvino_genai, m) {
 
     init_continuous_batching_pipeline(m);
     init_llm_pipeline(m);
-    init_text2image_pipeline(m);
+    init_image_generation_pipelines(m);
     init_vlm_pipeline(m);
     init_whisper_pipeline(m);
 }
